fixed order
by removing and replaying dice in both the cup and the table the issue is solved, as it was having issues with effectively dealing with six dice in play at any given time before sorting itself out. I also added the .reorder() functions to the beginning of the updateDice function to keep things tidy and looking proper.